# openapi.yaml
openapi: 3.0.3
info:
  title: Club Operations POS API
  version: 1.0.0
  description: |
    Server-authoritative REST API for managing club check-ins, room/locker assignments,
    cleaning workflows, checkout verification, and operational reporting.

    All endpoints are versioned under `/v1`.
    Request bodies are validated with Zod schemas.

    Authentication:
    - Staff endpoints require Bearer token (session token from PIN or WebAuthn login)
    - Admin endpoints require ADMIN role in addition to authentication
    - Customer-facing endpoints (agreement signing, checkout kiosk) do not require auth
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /v1/waitlist:
    get:
      tags: [Waitlist]
      summary: List waitlist entries (ACTIVE/OFFERED/etc)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, OFFERED, COMPLETED, CANCELLED, EXPIRED]
      responses:
        '200':
          description: Waitlist entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        desiredTier: { type: string, enum: [STANDARD, DOUBLE, SPECIAL] }
                        backupTier: { type: string }
                        status: { type: string }
                        createdAt: { type: string, format: date-time }
                        offeredAt: { type: string, format: date-time, nullable: true }
                        completedAt: { type: string, format: date-time, nullable: true }
                        displayIdentifier: { type: string }
                        currentRentalType: { type: string }
  /v1/waitlist/{id}/offer:
    post:
      tags: [Waitlist]
      summary: Offer upgrade to a waitlist entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [waitlistId, roomId]
              properties:
                waitlistId: { type: string, format: uuid }
                roomId: { type: string, format: uuid }
      responses:
        '200':
          description: Offered
  /v1/waitlist/{id}/cancel:
    post:
      tags: [Waitlist]
      summary: Cancel a waitlist entry (requires recent re-auth)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [waitlistId]
              properties:
                waitlistId: { type: string, format: uuid }
                reason: { type: string }
      responses:
        '200':
          description: Cancelled
        '403':
          description: Re-authentication required
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque session token from PIN or WebAuthn login (stored/validated server-side; not a JWT)
...